Extract shared DatabaseSchemaReader from ScaffoldCommand (step 1a)
Move CreateIntrospectorAsync + BuildConnectionString (and the per-table
introspection loop as ReadTablesAsync) out of ScaffoldCommand into a shared
Schema/DatabaseSchemaReader.cs, and promote TableIntrospectionData to a
top-level type. Refactor ScaffoldCommand to use the helper. Behavior-preserving
groundwork for migrate add --from-database / baseline / adopt.

Add DatabaseSchemaReader.cs to Quarry.Tests compile set and characterization
tests for BuildConnectionString across all four dialects.
