+++ src/Quarry.Tests/Quarry.Tests.csproj
@@ -75,6 +75,7 @@
   <ItemGroup>
     <Compile Include="../Quarry.Tool/Schema/ProjectSchemaReader.cs" LinkBase="Shared" />
     <Compile Include="../Quarry.Tool/Schema/DialectResolver.cs" LinkBase="Shared" />
+    <Compile Include="../Quarry.Tool/Schema/DatabaseSchemaReader.cs" LinkBase="Shared" />
     <Compile Include="../Quarry.Tool/Commands/BundleCommand.cs" LinkBase="Shared" />
     <Compile Include="../Quarry.Tool/Commands/CommandHelpers.cs" LinkBase="Shared" />
   </ItemGroup>
