+++ src/Quarry.Tool/Commands/ScaffoldCommand.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using Quarry.Shared.Migration;
 using Quarry.Shared.Scaffold;
 using Quarry.Tool.Interactive;
+using Quarry.Tool.Schema;
 
 namespace Quarry.Tool.Commands;
 
@@ -33,42 +34,24 @@ internal static class ScaffoldCommand
         var namingStyle = ParseNamingStyle(namingStyleStr);
 
         // Build connection string
-        var connStr = connectionString ?? BuildConnectionString(dialect, server, port, user, password, database);
+        var connStr = connectionString ?? DatabaseSchemaReader.BuildConnectionString(dialect, server, port, user, password, database);
         Console.WriteLine($"Connecting to {dialect} database...");
 
-        // Create introspector
-        using var introspector = await CreateIntrospectorAsync(dialect, connStr);
-
-        // Phase 1: Introspect tables
+        // Phase 1-3: Introspect tables (apply filter + gather full per-table metadata)
         Console.WriteLine("Introspecting database schema...");
-        var allTables = await introspector.GetTablesAsync(schemaFilter);
+        var introspectedTables = await DatabaseSchemaReader.ReadTablesAsync(dialect, connStr, schemaFilter, tables);
 
-        // Phase 2: Apply table filter
-        allTables = TableFilter.Apply(allTables, tables);
-
-        if (allTables.Count == 0)
+        if (introspectedTables.Count == 0)
         {
             Console.WriteLine("No tables found matching the specified criteria.");
             return;
         }
 
-        Console.WriteLine($"Found {allTables.Count} table(s).");
+        Console.WriteLine($"Found {introspectedTables.Count} table(s).");
 
-        // Log skipped views/non-table objects
-        // (filtered by introspector — only BASE TABLEs are returned)
-
-        // Phase 3: Gather full metadata for each table
         var tableData = new Dictionary<string, TableIntrospectionData>(StringComparer.OrdinalIgnoreCase);
-
-        foreach (var table in allTables)
-        {
-            var columns = await introspector.GetColumnsAsync(table.Name, table.Schema);
-            var pk = await introspector.GetPrimaryKeyAsync(table.Name, table.Schema);
-            var fks = await introspector.GetForeignKeysAsync(table.Name, table.Schema);
-            var indexes = await introspector.GetIndexesAsync(table.Name, table.Schema);
-
-            tableData[table.Name] = new TableIntrospectionData(table, columns, pk, fks, indexes);
-        }
+        foreach (var data in introspectedTables)
+            tableData[data.Table.Name] = data;
 
         // Build class name map
         var tableClassMap = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
@@ -297,82 +280,6 @@ internal static class ScaffoldCommand
         };
     }
 
-    private static async Task<IDatabaseIntrospector> CreateIntrospectorAsync(string dialect, string connectionString)
-    {
-        return dialect.ToLowerInvariant() switch
-        {
-            "sqlite" => await SqliteIntrospector.CreateAsync(connectionString),
-            "postgresql" or "postgres" or "pg" => await PostgreSqlIntrospector.CreateAsync(connectionString),
-            "sqlserver" or "mssql" => await SqlServerIntrospector.CreateAsync(connectionString),
-            "mysql" => await MySqlIntrospector.CreateAsync(connectionString),
-            _ => throw new InvalidOperationException($"Unknown dialect: {dialect}")
-        };
-    }
-
-    private static string BuildConnectionString(string dialect, string? server, string? port, string? user, string? password, string database)
-    {
-        return dialect.ToLowerInvariant() switch
-        {
-            "sqlite" => BuildSqliteConnectionString(database),
-            "postgresql" or "postgres" or "pg" => BuildNpgsqlConnectionString(server, port, user, password, database),
-            "sqlserver" or "mssql" => BuildSqlServerConnectionString(server, port, user, password, database),
-            "mysql" => BuildMySqlConnectionString(server, port, user, password, database),
-            _ => throw new InvalidOperationException($"Cannot build connection string for dialect: {dialect}")
-        };
-    }
-
-    private static string BuildSqliteConnectionString(string database)
-    {
-        var builder = new Microsoft.Data.Sqlite.SqliteConnectionStringBuilder { DataSource = database };
-        return builder.ConnectionString;
-    }
-
-    private static string BuildNpgsqlConnectionString(string? server, string? port, string? user, string? password, string database)
-    {
-        var builder = new Npgsql.NpgsqlConnectionStringBuilder
-        {
-            Host = server ?? "localhost",
-            Port = int.TryParse(port, out var p) ? p : 5432,
-            Database = database
-        };
-        if (user != null) builder.Username = user;
-        if (password != null) builder.Password = password;
-        return builder.ConnectionString;
-    }
-
-    private static string BuildSqlServerConnectionString(string? server, string? port, string? user, string? password, string database)
-    {
-        var builder = new Microsoft.Data.SqlClient.SqlConnectionStringBuilder
-        {
-            DataSource = (server ?? "localhost") + (port != null ? $",{port}" : ""),
-            InitialCatalog = database,
-            TrustServerCertificate = true
-        };
-        if (user != null)
-        {
-            builder.UserID = user;
-            if (password != null) builder.Password = password;
-        }
-        else
-        {
-            builder.IntegratedSecurity = true;
-        }
-        return builder.ConnectionString;
-    }
-
-    private static string BuildMySqlConnectionString(string? server, string? port, string? user, string? password, string database)
-    {
-        var builder = new MySqlConnector.MySqlConnectionStringBuilder
-        {
-            Server = server ?? "localhost",
-            Port = uint.TryParse(port, out var p) ? p : 3306,
-            Database = database
-        };
-        if (user != null) builder.UserID = user;
-        if (password != null) builder.Password = password;
-        return builder.ConnectionString;
-    }
-
     private static string ExtractDatabaseName(string database, string? connectionString)
     {
         if (!string.IsNullOrEmpty(database))
@@ -428,28 +335,5 @@ internal static class ScaffoldCommand
         }
 
         return current;
-    }
-
-    private sealed class TableIntrospectionData
-    {
-        public TableMetadata Table { get; }
-        public List<ColumnMetadata> Columns { get; }
-        public PrimaryKeyMetadata? PrimaryKey { get; }
-        public List<ForeignKeyMetadata> ForeignKeys { get; }
-        public List<IndexMetadata> Indexes { get; }
-
-        public TableIntrospectionData(
-            TableMetadata table,
-            List<ColumnMetadata> columns,
-            PrimaryKeyMetadata? primaryKey,
-            List<ForeignKeyMetadata> foreignKeys,
-            List<IndexMetadata> indexes)
-        {
-            Table = table;
-            Columns = columns;
-            PrimaryKey = primaryKey;
-            ForeignKeys = foreignKeys;
-            Indexes = indexes;
-        }
     }
 }
